Add feature_precision option to svm_train
It's sometimes necessary to specify the precision needed of the
features before sending to libsvm. By default, it assumes that
the in-place version wants singles, while the copy version takes
doubles. This isn't necessarily true, so we need an option for
this. In future versions of libsvm-compact, we can probably check
which precision is desired.