Fix inplace bug in svm_train
When calling svmtrain_inplace from svm_train, the whole set of
classes and instances must be given. However, this was only done
if a kernel was given. Otherwise, the subsampled set was provided.
To fix this bug, the whole set of instances is given in the non-
kernel case as well.